+++ src/Castle.Core/Core/Logging/ILogger.cs
@@ -27,6 +27,12 @@ namespace Castle.Core.Logging
 	public interface ILogger
 	{
 		/// <summary>
+		///   Determines if messages of priority "trace" will be logged.
+		/// </summary>
+		/// <value>True if "trace" messages will be logged.</value>
+		bool IsTraceEnabled { get; }
+
+		/// <summary>
 		///   Determines if messages of priority "debug" will be logged.
 		/// </summary>
 		/// <value>True if "debug" messages will be logged.</value>
@@ -64,6 +70,56 @@ namespace Castle.Core.Logging
 		/// <returns>The New ILogger instance.</returns>
 		/// <exception cref = "System.ArgumentException">If the name has an empty element name.</exception>
 		ILogger CreateChildLogger(String loggerName);
+
+		/// <summary>
+		///   Logs a trace message.
+		/// </summary>
+		/// <param name = "message">The message to log</param>
+		void Trace(String message);
+
+		/// <summary>
+		///   Logs a trace message with lazily constructed message. The message will be constructed only if the <see cref = "IsTraceEnabled" /> is true.
+		/// </summary>
+		void Trace(Func<string> messageFactory);
+
+		/// <summary>
+		///   Logs a trace message.
+		/// </summary>
+		/// <param name = "exception">The exception to log</param>
+		/// <param name = "message">The message to log</param>
+		void Trace(String message, Exception exception);
+
+		/// <summary>
+		///   Logs a trace message.
+		/// </summary>
+		/// <param name = "format">Format string for the message to log</param>
+		/// <param name = "args">Format arguments for the message to log</param>
+		void TraceFormat(String format, params Object[] args);
+
+		/// <summary>
+		///   Logs a trace message.
+		/// </summary>
+		/// <param name = "exception">The exception to log</param>
+		/// <param name = "format">Format string for the message to log</param>
+		/// <param name = "args">Format arguments for the message to log</param>
+		void TraceFormat(Exception exception, String format, params Object[] args);
+
+		/// <summary>
+		///   Logs a trace message.
+		/// </summary>
+		/// <param name = "formatProvider">The format provider to use</param>
+		/// <param name = "format">Format string for the message to log</param>
+		/// <param name = "args">Format arguments for the message to log</param>
+		void TraceFormat(IFormatProvider formatProvider, String format, params Object[] args);
+
+		/// <summary>
+		///   Logs a trace message.
+		/// </summary>
+		/// <param name = "exception">The exception to log</param>
+		/// <param name = "formatProvider">The format provider to use</param>
+		/// <param name = "format">Format string for the message to log</param>
+		/// <param name = "args">Format arguments for the message to log</param>
+		void TraceFormat(Exception exception, IFormatProvider formatProvider, String format, params Object[] args);
 
 		/// <summary>
 		///   Logs a debug message.